+++ R/zzz.R
@@ -9,7 +9,7 @@ R62S3::R62Fun(DistributionWrapper, assignEnvir = topenv(), dispatchClasses = lis
 
 .onAttach <- function(libname, pkgname) {
   packageStartupMessage("\n-------------------------------------------------")
-  packageStartupMessage(" \tWelcome to distr6 v",packageVersion("distr6"),
+  packageStartupMessage(" \tWelcome to distr6 v",utils::packageVersion("distr6"),
 "\n\t------------------------
 To get started see:
 * ?distr6