+++ NEWS.md
@@ -17,6 +17,7 @@
 - Added two additional parameters to `ProductDistribution`: `paramList` and `distribution`, so that it can either function as before or as the `ArrayDistribution` wrapper which is now deprecated.
 - Added two additional parameters to `VectorDistribution`: `paramList` and `distribution`, so that it can either function as before or as a generalisation to the `ArrayDistribution` wrapper which is now deprecated.
 - Removed automated rounding of numerics when `ParameterSet` expected integer in `setParameterValue`
+- Added parameter `n` to `strprint` and `print` to clean the print method for distributions (especially wrappers) with multiple parameters
 
 ## Patches